+++ expense_app/models.py
@@ -1,12 +1,5 @@
 from datetime import datetime
-from flask import Flask, render_template, flash, redirect, url_for
-from flask_sqlalchemy import SQLAlchemy
-from forms import RegistrationForm, LoginForm, ExpenseForm
-
-app = Flask(__name__)
-app.config['SECRET_KEY'] = '4fae35a28915d9bc651e0bc712350e5d'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
-db = SQLAlchemy(app)
+from expense_app import db
 
 
 class User(db.Model):
@@ -79,58 +72,3 @@ class PlannerItem(db.Model):
 
     def __repr__(self):
         return f"PlannerItem('{self.title}', '{self.planned_date}')"
-
-
-posts = [
-    {
-        'author': 'Corey Schafer',
-        'title': 'Blog Post 1',
-        'content': 'First post content',
-        'date_posted': 'April 20, 2018'
-
-    },
-    {
-        'author': 'Jane Doe',
-        'title': 'Blog Post 2',
-        'content': 'Second post content',
-        'date_posted': 'April 21, 2018'
-    }
-]
-
-
-@app.route("/")
-@app.route("/home")
-def home():
-    return render_template('home.html', posts=posts)
-
-
-@app.route("/about")
-def about():
-    return render_template('about.html', title='About')
-
-
-@app.route("/register", methods=['GET', 'POST'])
-def register():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        flash(f'Account created for {form.username.data}!', 'success')
-        return redirect(url_for('home'))
-    return render_template('register.html', title='Register', form=form)
-
-@app.route("/addexpense", methods=['GET', 'POST'])
-def addexpense():
-        form = ExpenseForm()
-        if form.validate_on_submit():
-            flash(f'Expense added!', 'success')
-            return redirect(url_for('home'))
-        return render_template('addExpense.html',title='Add Expense', form=form)
-
-
-@app.route("/login")
-def login():
-    form = LoginForm()
-    return render_template('login.html', title='Login', form=form)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
