+++ expense_app/templates/home.html
@@ -4,11 +4,11 @@
         <article class="media content-section">
           <div class="media-body">
             <div class="article-metadata">
-              <a class="mr-2" href="#">{{ expense.author }}</a>
-              <small class="text-muted">{{ expense.date_posted }}</small>
+              <a class="mr-2" href="#">{{ post.author }}</a>
+              <small class="text-muted">{{ post.date_posted }}</small>
             </div>
-            <h2><a class="article-title" href="#">{{ expense.title }}</a></h2>
-            <p class="article-content">{{ expense.content }}</p>
+            <h2><a class="article-title" href="#">{{ post.title }}</a></h2>
+            <p class="article-content">{{ post.content }}</p>
           </div>
         </article>
     {% endfor %}
