+++ run.py
@@ -1,0 +1,4 @@
+from expense_app import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
